Improve coding methodology and add png gen
The for loop to iterate every edges/states of an nfa has been updated
with map function. Trying to do some optimization on union operation and
kleene operation but still has bugs. The compile function returns 2
objects: one for my self-defined nfa structure, another for the object
that the fsm package generates.