Add optimization for node end edges
The kleene closure can be optimized as the same init and final state.
Add `maxId` to record the max number of states due to the optimization.
Add optimization to enable multiple final states. Redundent edges that
connects to a final state would be removed and 'from' state of the
redundent edge would be a new final state.